7. Reverse Integer
